Change JdbcUtil to resolve connection timeout issues with MySQL
The MySQL default is to timeout connections after 8 hour of
inactivity. The cBioPortal wiki recommends updating the
database connection Resource entry in the Tomcat server but
this is not available when running on Heroku or Flynn.io.

Rather, we try to do this in Java, and check the activity every 30
seconds, and force a timeout after 10 minutes of inactivity.